Move common code from pysph.base.opencl to cpy.
This further reduces any dependencies to pysph.  The opencl module here
simply makes it easy to get a queue/context etc.
